- Removed unsused stuff and improved stability

+++ client/app/account/settings/settings.controller.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('colorwatchApp')
-  .controller('SettingsCtrl', function ($scope, $http, User, Auth) {
+  .controller('SettingsCtrl', function ($scope, $http, User, Auth, $window) {
     $scope.errors = {};
 
     $scope.changePassword = function(form) {
@@ -9,11 +9,11 @@ angular.module('colorwatchApp')
       if(form.$valid) {
         Auth.changePassword( $scope.user.oldPassword, $scope.user.newPassword )
         .then( function() {
-          $scope.message = 'Password successfully changed.';
+          $scope.message = 'Lösenordet ändrat.';
         })
         .catch( function() {
           form.password.$setValidity('mongoose', false);
-          $scope.errors.other = 'Incorrect password';
+          $scope.errors.other = 'Fel lösenord';
           $scope.message = '';
         });
       }
@@ -26,17 +26,21 @@ angular.module('colorwatchApp')
 
     $scope.removeTest = function() {
       console.log('trying to remove test');
-      $http.delete('/api/results/').then(function(res){
-        console.log('Test removed', res);
-      }, function(err) {
-        console.log(err);
-      });
-      $http.delete('/api/colorcombs/').then(function(res) {
-        console.log('colorcombs removed', res);
-      }, function(err) {
-        console.log(err);
-      });
-      location.reload();
+
+      var removeTest = $window.confirm('Är du säker på att du vill radera hela testet och ta bort all insamlad data?');
+      if(removeTest) {
+        $http.delete('/api/results/').then(function(res){
+          console.log('Test removed', res);
+        }, function(err) {
+          console.log(err);
+        });
+        $http.delete('/api/colorcombs/').then(function(res) {
+          console.log('colorcombs removed', res);
+        }, function(err) {
+          console.log(err);
+        });
+        location.reload();
+      }
     };
     $scope.removeColor = function(colorId, index) {
       $http.delete('/api/colorcombs/' + colorId).then(function(){
